Update broken tests and outdated dependencies
- Dependencies `cypress-testing-library` and `react-testing-library` were
deprecated and replaced by `@testing-library/cypress` and
`@testing-library/react` respectively were the reason the
tests were failing. To fix the tests these libraries were updated.

- e2e tests which were using `waitForRouteChange` method were failing,
for some reason they were reaching a timeout error. That method was
replaced by the original wait on Cypress to avoid the timeout error; it
might better to try to find the issue on `waitForRouteChange`, but that
could take more time, maybe it could be open an issue on gatsby.